add column chart container
